prod/requirements.txt
alembic==1.12.1
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
Flask-WTF==1.2.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
PyJWT==2.8.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.1
